refactor/GETP229: ClientProjectListPage, ProjectListPage 폴더 정리
