+++ src/Router.tsx
@@ -3,8 +3,6 @@ import { Routes, Route } from "react-router-dom";
 import FindPasswordPage from "@/pages/auth/FindPasswordPage";
 import SignInPage from "@/pages/auth/SignInPage";
 import SignUpPage from "@/pages/auth/SignUpPage";
-import EditClientPage from "@/pages/client/EditClientPage";
-import RegisterClientPage from "@/pages/client/RegisterClientPage";
 import HomePage from "@/pages/home/HomePage";
 import NotFoundPage from "@/pages/home/NotFoundPage";
 import PeopleDetailPage from "@/pages/people/PeopleDetail/PeopleDetailPage";
@@ -13,7 +11,7 @@ import PeopleListPage from "@/pages/people/PeopleList/PeopleListPage";
 import PeopleProfileEditPage from "@/pages/people/PeopleProfileEdit/PeopleProfileEditPage";
 import ProjectApplyPage from "@/pages/project/ProjectApplyPage";
 import ProjectDetailPage from "@/pages/project/ProjectDetailPage";
-import ProjectListPage from "@/pages/project/ProjectListPage";
+import ProjectListPage from "@/pages/project/ProjectListPage/ProjectListPage";
 import ProjectRequestPage from "@/pages/project/ProjectRequestPage";
 
 import { MemberType } from "@/services/auth/types";
@@ -21,6 +19,8 @@ import { MemberType } from "@/services/auth/types";
 import { RouteGuard } from "./components/__common__/guard/RouteGuard";
 import { MainLayout } from "./components/__common__/layout/MainLayout";
 import ClientProjectListPage from "./pages/client/ClientProjectListPage/ClientProjectListPage";
+import EditClientPage from "./pages/client/EditClientPage";
+import RegisterClientPage from "./pages/client/RegisterClientPage";
 
 export const Router = () => {
     return (
